+++ core/CHANGELOG.md
@@ -419,9 +419,9 @@
 
 ## 0.24.0 [2020-11-09]
 
-- Remove `ConnectionInfo` trait and replace it with `PeerId`
+- Remove `Connection` trait and replace it with `PeerId`
   everywhere. This was already effectively the case because
-  `ConnectionInfo` was implemented on `PeerId`.
+  `Connection` was implemented on `PeerId`.
 
 ## 0.23.1 [2020-10-20]